Add compiled Khmer Sign Language landmark dataset
330 clips (11 signs x 30 clips), 90 frames x 258 landmark values,
saved as float32 (~29 MB). The train/val/test split files are
gitignored and regenerated by preprocess.py with a fixed seed;
results/split_info.json records the split sizes.

+++ src/extract_landmarks.py
@@ -60,7 +60,7 @@ def main():
         print("No valid sequences found. Check your data/ folder.")
         return
 
-    X = np.array(sequences)   # shape: (N, 30, 258)
+    X = np.array(sequences, dtype=np.float32)   # shape: (N, 90, 258); float32 halves the file size
     y = np.array(labels)       # shape: (N,)
 
     np.save(os.path.join(OUT_DIR, "X.npy"),      X)
